Add sha256
The `SHA256` hash is needed sometimes, however, because the `Solidity::hex()` method is private, the user doesn't have any means of reusing this great code with other hash types. 

Thus I would propose either of the following:

- make `hex()` public or
- add `sha256` public method

Up to you whether you want to merge it or not, but if you do, that would great. Any of the above two ways works for me. Making `hex()` public is probably a lot less intrusive, basically a one-liner edit.

Thx!